requirements.txt
python-telegram-bot==20.3
ffmpeg-python==0.2.0
httpx==0.24.1
telethon>=1.28.0
python-dotenv
